Refctoring step one: Get rid of heap allocations
Instead of allocating Nodes and Attributes on the free store those are
now stack allocated and moved around. In a few cases there might still
be unneccisary copies, but it should be simpler and fixes at least two
potential leaks.

With this patch the data of an attribute is stored in a fixed array,
this shoul be big enough for most data except the timtable XML, however
the current state of network handling only reads a single packet, which
can lead to patial messages.

Next steps are some restructuring of network code (i.e. Attributes and
Node aren't responsble for sending data themselves) and adding a nicer
abstraction on the attributes, that a consumer doesn't have to know
each fields representation but can ask the library for the type and
handling of the data.